+++ Example/Pods/Target Support Files/Pods-RecastAI_Tests/Pods-RecastAI_Tests-frameworks.sh
@@ -84,16 +84,18 @@ strip_invalid_archs() {
 
 
 if [[ "$CONFIGURATION" == "Debug" ]]; then
+  install_framework "Pods-RecastAI_Tests/Alamofire.framework"
   install_framework "Pods-RecastAI_Tests/JSONJoy.framework"
   install_framework "Pods-RecastAI_Tests/RecastAI.framework"
-  install_framework "Pods-RecastAI_Tests/SwiftHTTP.framework"
+  install_framework "Pods-RecastAI_Tests/SwiftWebSocket.framework"
   install_framework "Pods-RecastAI_Tests/Nimble.framework"
   install_framework "Pods-RecastAI_Tests/Quick.framework"
 fi
 if [[ "$CONFIGURATION" == "Release" ]]; then
+  install_framework "Pods-RecastAI_Tests/Alamofire.framework"
   install_framework "Pods-RecastAI_Tests/JSONJoy.framework"
   install_framework "Pods-RecastAI_Tests/RecastAI.framework"
-  install_framework "Pods-RecastAI_Tests/SwiftHTTP.framework"
+  install_framework "Pods-RecastAI_Tests/SwiftWebSocket.framework"
   install_framework "Pods-RecastAI_Tests/Nimble.framework"
   install_framework "Pods-RecastAI_Tests/Quick.framework"
 fi