+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 ---
+## [1.0.0](https://github.com/RecastAI/SDK-ios/releases/tag/1.0.0)
+
+#### Features
+
+* Add Voice
+* Change pod to Alamofire for HTTP Request
+* Add voice to example project
+
 ## [0.6.0](https://github.com/RecastAI/SDK-ios/releases/tag/0.6.0)
 
 #### Features
@@ -18,5 +26,5 @@ All notable changes to this project will be documented in this file.
 
 #### Features
 
-* SDK Swfit
+* SDK Swift
 * Update README